Создал класс DBModel, унаследованный от Model, и содержащий методы для реализации CRUD-операций
