+++ models/Cart.php
@@ -2,7 +2,7 @@
 
 namespace app\models;
 
-class Cart extends Model
+class Cart extends DBModel
 {
 
     public $id;
