+++ models/Product.php
@@ -2,7 +2,7 @@
 
 namespace app\models;
 
-class Product extends Model {
+class Product extends DBModel {
     public $id;
     public $name_product;
     public $price;
