+++ public/index.php
@@ -6,13 +6,13 @@ use app\engine\{Autoload, Db};
 include "../engine/Autoload.php";
 spl_autoload_register([new Autoload, "loadClass"]);
 
-// $product = Product::getOne(3);
+$product = Product::getOne(3);
+var_dump($product);
+
+// $product = new Product('test_prod22', 10000, 'test22.jpg', 4, 5);
 // var_dump($product);
 
-$product = new Product('test_prod22', 10000, 'test22.jpg', 4, 5);
-// var_dump($product);
-
-$product->insert();
+// $product->insert();
 // var_dump($product);
 
 // if ($product->delete()) unset($product);
